AICodingWeb/aicoding_web/requirements.txt
# Core dependencies
Werkzeug==2.3.7
Jinja2==3.1.2
click==8.1.7
itsdangerous==2.1.2
MarkupSafe==2.1.3

# Flask and extensions
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.2
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Flask-Babel==4.0.0

# Database
#pyodbc==4.0.39
SQLAlchemy==2.0.23

# Security
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT==2.8.0

# Azure Key Vault (for production)
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
email-validator==2.1.0.post1

# Development tools
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0 